requirements.txt
fastapi
uvicorn
pymongo
python-multipart
python-dotenv
insightface
numpy
onnx
onnxruntime
opencv-python #not needed since headless is used.
opencv-python-headless
imutils
boto3
jinja2
celery
redis
flower
pyjwt
passlib
bcrypt

